+++ src/index.tsx
@@ -9,7 +9,7 @@ import { GenericListing } from "routes/GenericListing";
 import { TomogramPage } from "routes/Tomogram";
 import { SpaPage } from "routes/SPA";
 import { Error } from "routes/Error";
-import { Accordion, Button, Text, Heading, Table, Card, Tabs, Checkbox } from "styles/components";
+import { Accordion, Button, Text, Heading, Table, Card, Tabs, Checkbox, Code } from "styles/components";
 import { colours } from "styles/colours";
 import { Home } from "routes/Home";
 import {
@@ -30,14 +30,14 @@ const { ToastContainer } = createStandaloneToast();
 const container = document.getElementById("root")!;
 const root = createRoot(container);
 
-if (process.env.REACT_APP_DEMO === 'true') {
-  const { worker } = require('./mocks/browser')
-  worker.start()
+if (process.env.REACT_APP_DEMO === "true") {
+  const { worker } = require("./mocks/browser");
+  worker.start();
 }
 
 const theme = extendTheme({
   colors: colours,
-  components: { Accordion, Checkbox, Button, Text, Heading, Table, Card, Tabs },
+  components: { Accordion, Checkbox, Button, Text, Heading, Table, Card, Tabs, Code },
   breakpoints: {
     "sm": "30em",
     "md": "48em",
@@ -47,7 +47,9 @@ const theme = extendTheme({
   },
 });
 
-const checkUrlChanged = (current: URL, next: URL) => current.href !== next.href;
+const checkListingChanged = (current: URL, next: URL) =>
+  (current.searchParams.get("items") !== null || current.searchParams.get("page") !== null) &&
+  current.href !== next.href;
 
 const handleGroupClicked = (item: Record<string, string | number>) => {
   // Temporary workaround
@@ -102,7 +104,7 @@ const router = createBrowserRouter([
           />
         ),
         loader: ({ request, params }) => getListingData(request, params, "proposals"),
-        shouldRevalidate: ({ currentUrl, nextUrl }) => checkUrlChanged(currentUrl, nextUrl),
+        shouldRevalidate: ({ currentUrl, nextUrl }) => checkListingChanged(currentUrl, nextUrl),
       },
       {
         path: "/calendar",
@@ -126,7 +128,7 @@ const router = createBrowserRouter([
           />
         ),
         loader: ({ request, params }) => getListingData(request, params, "sessions", processSessionData),
-        shouldRevalidate: ({ currentUrl, nextUrl }) => checkUrlChanged(currentUrl, nextUrl),
+        shouldRevalidate: ({ currentUrl, nextUrl }) => checkListingChanged(currentUrl, nextUrl),
       },
       {
         path: "/proposals/:propid/sessions/:visitId",
@@ -153,7 +155,7 @@ const router = createBrowserRouter([
           />
         ),
         loader: ({ request, params }) => getListingData(request, params, "dataCollections"),
-        shouldRevalidate: ({ currentUrl, nextUrl }) => checkUrlChanged(currentUrl, nextUrl),
+        shouldRevalidate: ({ currentUrl, nextUrl }) => checkListingChanged(currentUrl, nextUrl),
       },
       {
         path: "/proposals/:propId/sessions/:visitId/groups/:groupId/tomograms/",